fixed davclient for iCloud.

+++ www/go/modules/community/davclient/model/DavAccount.php
@@ -103,10 +103,12 @@ class DavAccount extends AclOwnerEntity {
 
 		$ok = parent::internalDelete($query);
 		if($ok && !self::$keepData && \go\core\Module::isAllowed('calendar', 'community')) {
-			// remove the calendars after
-			if(!\go\modules\community\calendar\model\Calendar::delete(['id'=>$calIDs])) {
-				$ok = false;
-				throw new \Exception("Unable to delete calendars related to dav account");
+			if(!empty($calIDs)) {
+				// remove the calendars after
+				if (!\go\modules\community\calendar\model\Calendar::delete(['id' => $calIDs])) {
+					$ok = false;
+					throw new \Exception("Unable to delete calendars related to dav account");
+				}
 			}
 
 		}
@@ -136,8 +138,21 @@ class DavAccount extends AclOwnerEntity {
 		}
 
 		if (!isset($host)) {
+			// Thunderbird will do HEAD /, GET /, PROPFIND .well-known
 			$data = $this->http()->get("/.well-known/$this->service");
-			$this->basePath = parse_url(rtrim($data['headers']['location'],'/'), PHP_URL_PATH).'/';
+			if(isset($data['headers']['location'])) {
+				$this->basePath = parse_url(rtrim($data['headers']['location'], '/'), PHP_URL_PATH) . '/';
+				$this->principalUri = $this->principalUri();
+			} else {
+				$responses = $this->propfind(['d:current-user-principal'], "/.well-known/$this->service");
+				foreach ($responses as $href => $response) {
+					if (isset($response->{'current-user-principal'})) {
+						$this->principalUri = $href;
+						return;
+					}
+				}
+				throw new \Exception("Could not find principalUri");
+			}
 		}
 	}
 
@@ -145,7 +160,7 @@ class DavAccount extends AclOwnerEntity {
 	private function serviceDiscovery() {
 
 		$this->dnsResolve(); // or well-known
-		$this->principalUri = $this->principalUri();
+
 
 		// todo: remove if double
 		$this->lastSync = new \DateTime();
@@ -164,7 +179,7 @@ class DavAccount extends AclOwnerEntity {
 //		}
 		$homesetUri = $this->homeSetUri($this->principalUri);
 
-		go()->getDbConnection()->beginTransaction();
+
 
 		$responses = $this->syncCollections($homesetUri);
 
@@ -185,9 +200,12 @@ class DavAccount extends AclOwnerEntity {
 				$collection = $responses[$calendar->uri];
 				if($calendar->isNew() || $calendar->ctag !== (string)$collection->getctag) {
 					// resync
+					go()->getDbConnection()->beginTransaction();
+					go()->log('Synchronizing '. $calendar->uri. ' ctag mismatch ['. $calendar->ctag .' != '. $collection->getctag.']');
 					if($calendar->sync()) {
 						$calendar->ctag = (string)$collection->getctag;
 					}
+					go()->getDbConnection()->commit();
 				}
 				unset($responses[$calendar->uri]);
 			}
@@ -195,12 +213,11 @@ class DavAccount extends AclOwnerEntity {
 
 		$this->lastSync = new \DateTime();
 		if(!$this->save()) {
-			go()->log('Could not save dav account '. $homesetUri);
-			go()->getDbConnection()->rollBack();
+			go()->log('Could not save last sync '. $homesetUri);
+			//go()->getDbConnection()->rollBack();
 		} else if(!empty($deletedCalendars)) {
 			\go\modules\community\calendar\model\Calendar::delete((new Query())->where('id','IN',$deletedCalendars));
 		}
-		go()->getDbConnection()->commit();
 		// fetch
 	}
 
@@ -217,22 +234,28 @@ class DavAccount extends AclOwnerEntity {
 			'cs:getctag'
 		], $homesetUri, 1);
 
+		$calendars = [];
+
 		foreach ($responses as $href => $response) {
 //			if (isset($response->resourcetype->addressbook)) {
 //				$this->addAddressbook($href, $response);
 //			}
 			$isCalendar = false;
-			if(isset($response->{'supported-calendar-component-set'})) {
-				$isCalendar = (string)$response->{'supported-calendar-component-set'}->comp->attributes()->name === 'VEVENT';
+			if(isset($response->resourcetype->calendar)) {
+				$isCalendar = true;
+				if(isset($response->{'supported-calendar-component-set'})) {
+					$isCalendar = (string)$response->{'supported-calendar-component-set'}->comp->attributes()->name === 'VEVENT';
+				}
 			}
 
 			if ($isCalendar) {
 				$this->addCalendar($href, $response);
+				$calendars[$href] = $response;
 			} else {
 				// $this->addTasklist()??
 			}
 		}
-		return $responses;
+		return $calendars;
 	}
 
 	/**
@@ -266,7 +289,7 @@ class DavAccount extends AclOwnerEntity {
 				$cal->id = $model->id;
 				$cal->uri = $uri;
 				$cal->ctag = '';// (string) $response->getctag;
-				$cal->synctoken = (string) $response->{'sync-token'};
+				$cal->synctoken = '' ;// (string) $response->{'sync-token'}; ( not supported at the moment)
 				$this->collections[$cal->id] = $cal;
 				//$cal->sync();
 			} else {
